♻️ Refactor: 체험 수정 관련 타입을 도메인 내부로 이동
