+++ src/features/my-page/activity-form/hooks/useActivityEditSubmit.ts
@@ -1,9 +1,9 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
-import type { CreateActivityResponse } from '@/features/activity/types';
 import { useUpdateActivity } from '@/features/my-page/activity-form/mutations/useUpdateActivity';
 import { useUploadImage } from '@/features/my-page/activity-form/mutations/useUploadImage';
+import type { CreateActivityResponse } from '@/features/my-page/activity-form/types';
 import type { ActivityFormValues } from '@/features/my-page/activity-form/utils/schema';
 import { QUERY_KEYS } from '@/shared/constants/queryKey';
 import { useToastStore } from '@/shared/ui/toast/stores/useToastStore';
