+++ src/features/my-page/types.ts
@@ -161,62 +161,3 @@ export type UpdateMyActivityReservationStatusResponse = z.infer<
 // ─── 내 체험 삭제 ───────────────────────────────────────────────
 
 export type DeleteMyActivityResponse = undefined;
-
-/** 체험 수정 요청 */
-export const UpdateMyActivityRequestBodySchema = z.object({
-  title: z.string(),
-  category: z.string(),
-  description: z.string(),
-  price: z.number(),
-  address: z.string(),
-  bannerImageUrl: z.string(),
-  subImageIdsToRemove: z.array(z.number()),
-  subImageUrlsToAdd: z.array(z.string()),
-  scheduleIdsToRemove: z.array(z.number()),
-  schedulesToAdd: z.array(
-    z.object({
-      date: z.string(),
-      startTime: z.string(),
-      endTime: z.string(),
-    })
-  ),
-});
-
-/** 체험 수정 응답 내 스케줄 */
-export const UpdateMyActivityScheduleSchema = z.object({
-  date: z.string(),
-  times: z.array(
-    z.object({
-      id: z.number(),
-      startTime: z.string(),
-      endTime: z.string(),
-    })
-  ),
-});
-
-/** 체험 수정 응답 내 서브 이미지 */
-export const UpdateMyActivitySubImageSchema = z.object({
-  id: z.number(),
-  imageUrl: z.string(),
-});
-
-/** 체험 수정 응답 */
-export const UpdateMyActivityResponseSchema = z.object({
-  id: z.number(),
-  userId: z.number(),
-  title: z.string(),
-  description: z.string(),
-  category: z.string(),
-  price: z.number(),
-  address: z.string(),
-  bannerImageUrl: z.string(),
-  rating: z.number(),
-  reviewCount: z.number(),
-  createdAt: z.string(),
-  updatedAt: z.string(),
-  subImages: z.array(UpdateMyActivitySubImageSchema),
-  schedules: z.array(UpdateMyActivityScheduleSchema),
-});
-
-export type UpdateMyActivityRequestBody = z.infer<typeof UpdateMyActivityRequestBodySchema>;
-export type UpdateMyActivityResponse = z.infer<typeof UpdateMyActivityResponseSchema>;
